Add files via upload
for testing PCB after soldering. i dont know if the code wil work. the adafruit_NeoMatrix lib does color in a diffrent way because its originally for lcd wich use a 565 rgb value and we have a 888rgb value and also the white so a rgbw8888 value. so i hope it works. i also dont know how fast it is. and brightness maybe very low because i definde the brightness 50/255 and the 565 rgb value is only the lowest 2 bits of each color is high. hopefully it works